+++ e2e/libs/utility/utility.py
@@ -15,9 +15,6 @@ from kubernetes.client.rest import ApiException
 from robot.api import logger
 from robot.libraries.BuiltIn import BuiltIn
 
-from node.utility import get_node_by_index
-from node.utility import list_node_names_by_role
-
 
 def logging(msg, also_report=False):
     if also_report:
@@ -72,27 +69,6 @@ def wait_for_cluster_ready():
             logging(f"Listing nodes error: {e}")
         time.sleep(retry_interval)
     assert ready, f"expect cluster's ready but it isn't {resp}"
-
-
-def wait_for_all_instance_manager_running():
-    longhorn_client = get_longhorn_client()
-    worker_nodes = list_node_names_by_role("worker")
-
-    retry_count, retry_interval = get_retry_count_and_interval()
-    for _ in range(retry_count):
-        logging(f"Waiting for all instance manager running ({_}) ...")
-        try:
-            instance_managers = longhorn_client.list_instance_manager()
-            instance_manager_map = {}
-            for im in instance_managers:
-                if im.currentState == "running":
-                    instance_manager_map[im.nodeID] = im
-            if len(instance_manager_map) == len(worker_nodes):
-                break
-            time.sleep(retry_interval)
-        except Exception as e:
-            logging(f"Getting instance manager state error: {e}")
-    assert len(instance_manager_map) == len(worker_nodes), f"expect all instance managers running, instance_managers = {instance_managers}, instance_manager_map = {instance_manager_map}"
 
 
 def apply_cr(manifest_dict):
@@ -196,21 +172,6 @@ def get_longhorn_client():
             except Exception as e:
                 logging(f"Getting longhorn client error: {e}")
                 time.sleep(retry_interval)
-
-
-def get_test_pod_running_node():
-    if "NODE_NAME" in os.environ:
-        return os.environ["NODE_NAME"]
-    else:
-        return get_node_by_index(0)
-
-
-def get_test_pod_not_running_node():
-    worker_nodes = list_node_names_by_role("worker")
-    test_pod_running_node = get_test_pod_running_node()
-    for worker_node in worker_nodes:
-        if worker_node != test_pod_running_node:
-            return worker_node
 
 
 def get_test_case_namespace(test_name):
